refactor: removed answerRestrictions file
